+++ examples/models/tensorRTServing/s2i/mnist_transformer/environment
@@ -1,0 +1,4 @@
+MODEL_NAME=MnistTransformer
+API_TYPE=REST
+SERVICE_TYPE=TRANSFORMER
+PERSISTENCE=0
