+++ examples/models/tensorRTServing/s2i/tensorrt-proxy/environment
@@ -1,0 +1,4 @@
+MODEL_NAME=TRTProxy
+API_TYPE=REST
+SERVICE_TYPE=MODEL
+PERSISTENCE=0
