+++ wrappers/s2i/python-tensorrt/Makefile
@@ -1,5 +1,5 @@
 IMAGE_VERSION=0.1
-IMAGE_NAME = docker.io/seldonio/seldon-core-s2i-python3-ngraph-onnx
+IMAGE_NAME = docker.io/seldonio/seldon-core-s2i-python3-tensorrt
 
 SELDON_CORE_DIR=../../..
 
@@ -9,3 +9,4 @@ build:
 
 push_to_dockerhub:
 	docker push $(IMAGE_NAME):$(IMAGE_VERSION)
+
